+++ src/pool/pool.ts
@@ -61,7 +61,7 @@ class Pool {
     }
 
     /**
-     * The current number of item in the pool's queue.
+     * The current number of item in the pool's queue on the current thread.
      */
     get queueLength() {
         return this.#queue.length;
@@ -114,7 +114,7 @@ class Pool {
      */
     __enqueue(item: PoolItem) {
         // Prevent workers from being run on any other thread than the main thread.
-        if (!isMainThread) throw new Error("Can't enqueue items to the pool on any other thread than the main thread!");
+        // if (!isMainThread) throw new Error("Can't enqueue items to the pool on any other thread than the main thread!");
         if (!(item instanceof PoolItem)) throw new Error('The provided item cannot be enqueued.');
 
         if (item.options.priority) this.#queue.unshift(item);