+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.4.1] - 2023-5-2
+
 ### Changed
 
 - `pool` to use shared values for the active count and concurrency count, enabling the spawning of new threads from other threads.